認証機能の作成

+++ resources/views/show.blade.php
@@ -10,11 +10,15 @@
       width='50%' height='300' frameborder='0'>
     </iframe>
   </div>
-  <div>
-    <a href = '{{ route("shop.edit", ["shop" => $shop->id]) }}', class='btn btn-primary'>編集する</a>
-    <a href = '{{ route("shop.list") }}'>一覧にもどる</a>
-    {{ Form::open(['method' => 'delete', 'route' => ['shop.destroy', $shop->id] ]) }}
-      {{ Form::submit('削除', ['class' => 'btn btn-outline-danger']) }}
-    {{ Form::close() }}
+  <div class="d-flex">
+    @auth   <!-- 投稿者にのみ、編集ボタン、削除ボタンを表示 -->
+      @if ($shop -> user_id == $login_user_id)
+        <a href = '{{ route("shop.edit", ["shop" => $shop->id]) }}', class='btn btn-primary mx-3'>編集</a>
+        {{ Form::open(['method' => 'delete', 'route' => ['shop.destroy', $shop->id] ]) }}
+          {{ Form::submit('削除', ['class' => 'btn btn-outline-danger']) }}
+        {{ Form::close() }}
+      @endif
+    @endauth
+    <a href = '{{ route("shop.list") }}' class="d-flex align-items-center btn btn-outline-secondary mx-3">一覧にもどる</a>
   </div>
 @endsection
